New commit update

+++ database/migrations/2024_11_20_065608_create_butiran_pengusaha.php
@@ -11,27 +11,29 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('butiran_pengusahas', function (Blueprint $table) {
-            $table->id();
-            // $table->string('applicant_id')->nullable();
-            $table->unsignedBigInteger('applicant_id')->nullable();
-            $table->foreign('applicant_id')->references('id')->on('applicants')->onDelete('cascade');
-            
+        if (!Schema::hasTable('butiran_pengusahas')) {
+            Schema::create('butiran_pengusahas', function (Blueprint $table) {
+                $table->id();
+                // $table->string('applicant_id')->nullable();
+                $table->unsignedBigInteger('applicant_id')->nullable();
+                $table->foreign('applicant_id')->references('id')->on('applicants')->onDelete('cascade');
+                
 
-            //2.BUTIRAN PENGUSAHA (A. Sekiranya individu)
-            $table->string('nama_pengusaha')->nullable();
-            $table->string('kad_pengenalan_pengusaha')->nullable();
-            $table->string('warganegara_pengusaha')->nullable();
-            $table->longText('alamat_rumah_pengusaha')->nullable();
-            $table->longText('alamat_berlainan_pengusaha')->nullable();
-            $table->string('email_pengusaha')->nullable();
-            $table->string('telefon_pengusaha')->nullable();
-            $table->string('faks_pengusaha')->nullable();
-            $table->string('jenis_pengusaha')->nullable();
-           
-            $table->timestamps();
-            
-        });
+                //2.BUTIRAN PENGUSAHA (A. Sekiranya individu)
+                $table->string('nama_pengusaha')->nullable();
+                $table->string('kad_pengenalan_pengusaha')->nullable();
+                $table->string('warganegara_pengusaha')->nullable();
+                $table->longText('alamat_rumah_pengusaha')->nullable();
+                $table->longText('alamat_berlainan_pengusaha')->nullable();
+                $table->string('email_pengusaha')->nullable();
+                $table->string('telefon_pengusaha')->nullable();
+                $table->string('faks_pengusaha')->nullable();
+                $table->string('jenis_pengusaha')->nullable();
+               
+                $table->timestamps();
+                
+            });
+        }
     }
 
     /**
